Migrate from FixedBitSet to RoaringBitmaps

+++ src/lib.rs
@@ -3,6 +3,6 @@ pub mod digraph;
 pub mod strictly_upper_triangular_logical_matrix;
 
 trait TraversableDirectedGraph {
-    fn extend_with_children(&self, children: &mut Vec<usize>, u: usize);
-    fn extend_with_parents(&self, parents: &mut Vec<usize>, v: usize);
+    fn extend_with_children(&self, children: &mut Vec<u32>, u: u32);
+    fn extend_with_parents(&self, parents: &mut Vec<u32>, v: u32);
 }
